requirements.txt
requests==2.32.4
python-can==4.6.1
pyyaml==6.0.2
python-dotenv==1.1.1
cantools==40.5.0